% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_stamp}
\alias{add_stamp}
\title{Generate footnotes - including program name, timestamp, and data source attached.}
\usage{
add_stamp(fnote_list = footnote_list, ...)
}
\arguments{
\item{df}{A dataframe or list of titles and footnotes.}

\item{tnumber}{TFL number, used to select proper titles and footnotes. This can be with or without TFL type.If pname parameter is not given, tnumber must not be NULL.}

\item{type}{optional TFL type}

\item{pname}{The program name used to select proper title entry. If this parameter is given (not NULL), it has precedence over TFL number for selection.}

\item{oid}{Optional parameter}
}
\description{
This function prepares footnotes for tables and listings, or figures.
}
\details{
Generate a list of footnotes with an additional line of program name, run timestamp, and data source(s) added to the end.
}
