% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_footer_for_flextable.R
\name{footer_for_flextable}
\alias{footer_for_flextable}
\title{Generate footnotes - including program name, timestamp, and data source attached.}
\usage{
footer_for_flextable(filename, type = NA, tnumber = NA, pname = NA, oid = NA)
}
\arguments{
\item{filename}{Full filename including folder path and file extension.}
}
\description{
This function prepares footnotes for tables and listings using {flextable} package.
}
\details{
Calls other utils functions.
}
