% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_refstamp}
\alias{get_refstamp}
\title{Generate Runtime Program Name Reference}
\usage{
get_refstamp(pname = NULL)
}
\arguments{
\item{pname}{Optional. A program name to include in the timestamp message.}
}
\value{
A character string with the reference program name.
}
\description{
Generates a string indicating the program name from which the report was generated.
}
