% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_source}
\alias{get_source}
\title{Return data source(s)}
\usage{
get_source(df = NULL, type = NULL, tnumber = NULL, pname = NULL, oid = NULL)
}
\arguments{
\item{df}{Full filename including folder path and file extension.}

\item{type}{optional TFL type}

\item{tnumber}{TFL number, used to select proper titles and footnotes. This can be with or without TFL type.If pname parameter is not given, tnumber must not be NULL.}

\item{pname}{The program name used to select proper title entry. If this parameter is given (not NULL), it has precedence over TFL number for selection.}

\item{oid}{Optional parameter}
}
\description{
This function assumes that the data source(s) along with the titles and footnotes are stored in a spreadsheet
}
\details{
Calls other utils functions.
}
