% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_timestamp}
\alias{get_timestamp}
\title{This function returns program name and timestamp for printing at the bottom of TFLs}
\usage{
get_timestamp(pname = NULL)
}
\description{
This function returns program name and timestamp for printing at the bottom of TFLs
}
\details{
Calls other utils functions.
}
