% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_footer_for_flextable.R
\name{header_for_flextable}
\alias{header_for_flextable}
\title{Generate headers - titles and population}
\usage{
header_for_flextable(filename, type = NA, tnumber = NA, pname = NA, oid = NA)
}
\arguments{
\item{filename}{Full filename including folder path and file extension.}
}
\description{
This function reads headers for tables and listings using {flextable} package.
}
\details{
Calls other utils functions. The header should be unlisted.
}
