% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_footer}
\alias{read_footer}
\title{Read header and footer Excel file}
\usage{
read_footer(filename = filename)
}
\description{
This function reads headers and footers from a spreadsheet, either an Excel file or csv file.
}
\details{
Table_layout is based upon the complex_layout. However, unlike complex_layout, the table layout consists of eight rows for headers, titles, plot, notes, and footnotes.
The fourth row and sixth row are used to create space above and below the table.\cr
The heights of the rows in table_layout with \code{"free"} scales are 5\%, 5\%, 5\%, 5\%, 60\%, 5\%, 5\%, and 10\% of the area respectively.\cr
In \code{table_layout} with \code{"fixed"} scales, row heights are specified in inches for annotations, while the remaining space is dedicated to the plot.
This ensures consistent spacing across outputs, with the plot occupying the central area.\cr
Please note that as output space is reduced, annotations retain their space which makes the plot appear smaller.

For an example, see \code{vignette("table_example", package = "plotation")}.
}
\examples{
table_layout()

# (to use |> version 4.1.0 of R is required, for lower versions we recommend \%>\% from magrittr)
library(magrittr)

library(flextable)



}
\seealso{
\link{read_header}
}
