% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_xlfile}
\alias{read_xlfile}
\title{Read and Validate Excel File}
\usage{
read_xlfile(filename, sheetname)
}
\arguments{
\item{filename}{A string. The path to the Excel file.}

\item{sheetname}{A string. The name of the sheet within the Excel file to read.}
}
\value{
A data frame containing the contents of the Excel sheet, if all required columns
(\code{PGMNAME}, \code{TTL1}, \code{FOOT1}, and \code{SOURCE}) are present. Otherwise, the function throws an error.
}
\description{
Reads an Excel file and verifies that it contains all required columns.
Column names are converted to uppercase before validation.
}
\details{
This function reads an Excel file using \code{\link[readxl:read_excel]{readxl::read_excel()}} and ensures the
presence of specific required columns. It is intended to standardize input structure
for downstream processing of metadata related to clinical tables and figures.
}
\examples{
\dontrun{
read_xlfile("metadata.xlsx", "Sheet1")
}

}
