% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_with_name}
\alias{select_with_name}
\title{Select header and footer based on program name and oid}
\usage{
select_with_name(df = list(), pname = "", oid = "")
}
\arguments{
\item{df}{A dataframe or list of overall title and footnote}

\item{pname}{The program name used to select proper title entry. If this parameter is given (not NA), it has precedence over TFL number for selection.}

\item{oid}{Optional parameter}
}
\description{
This function select titles and footnotes based on TFL number
}
\details{
Each TFL (table, listing, or figure) should have a type and TFL number in the header_footer Excel spreadsheet. This information can be in two separate columns (TYPE & TTL1)
or combined in TTL1 column. However, if both type and TFL number are already combined in column TTL1, there should be a space to separate them.
Additional details...
}
