% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{change_colname}
\alias{change_colname}
\title{Change column names for titles and footnotes etc.}
\usage{
change_colname(
  df,
  title_name = NULL,
  ftnote_name = NULL,
  pg_name = NULL,
  oid_name = NULL,
  pop_name = NULL,
  source_name = NULL,
  byline_name = NULL,
  bookm_name = NULL,
  type_name = NULL,
  ...
)
}
\arguments{
\item{df}{A data frame whose column names are to be changed.}

\item{title_name}{Character vector of columns to rename as titles (\code{TTL1}, \code{TTL2}, ...).}

\item{ftnote_name}{Character vector of columns to rename as footnotes (\code{FOOT1}, \code{FOOT2}, ...).}

\item{pg_name}{Character vector of columns to rename as program name (\code{Pgmname}).}

\item{oid_name}{Character vector of columns to rename as output OIDs (\code{OID}).}

\item{pop_name}{Character vector of columns to rename as population (\code{Population}).}

\item{source_name}{Character vector of columns to rename as dataset sources (\code{Source}).}

\item{byline_name}{Character vector of columns to rename as bylines (\code{BYLINE1}, \code{BYLINE2}, ...).}

\item{bookm_name}{Character vector of columns to rename as bookmark (\code{BOOKM}).}

\item{type_name}{Character vector of columns to rename as type (\code{Type}).}

\item{...}{Additional named mappings where the name is the new prefix and the
value is a character vector of existing column names.}
}
\value{
A data frame with updated column names.
}
\description{
This function renames user-specified columns in a data frame.
Title columns are renamed sequentially as \code{TTL1}, \code{TTL2}, …
Footnote columns are renamed sequentially as \code{FOOT1}, \code{FOOT2}, …
Byline columns are renamed sequentially as \code{BYLINE1}, \code{BYLINE2}, …
}
\examples{
df <- data.frame(
  TitleA = 1:3,
  FootnoteA = 4:6,
  PageX = 7:9,
  PopX = 10:12,
  Data = 13:15
)

change_colname(df,
  title_name = "TitleA",
  ftnote_name = "FootnoteA",
  pg_name = "PageX",
  pop_name = "PopX"
)

# Mixing both
change_colname(df,
  title_names = "TitleA",
  TFLTYPE = "Type"
)

}
