% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_byline}
\alias{get_byline}
\title{Extract Byline Metadata}
\usage{
get_byline(df = NULL, type = NULL, tnumber = NULL, pname = NULL, oid = NULL)
}
\arguments{
\item{df}{A data frame of title and footnote metadata.}

\item{type}{Optional. A character string indicating the TFL type.}

\item{tnumber}{Optional. TFL number.}

\item{pname}{Optional. Program name.}

\item{oid}{Optional. Object ID.}
}
\value{
A named list of non-missing byline fields.
}
\description{
Retrieves byline fields (BYLINE1, BYLINE2, etc.) from the metadata.
}
