% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_footnote}
\alias{get_footnote}
\title{Extract Footnote Metadata}
\usage{
get_footnote(
  df = NULL,
  type = NULL,
  tnumber = NULL,
  pname = NULL,
  oid = NULL,
  add_footr_tstamp = TRUE
)
}
\arguments{
\item{df}{A data frame of title and footnote metadata.}

\item{type}{Optional. A character string indicating the TFL type.}

\item{tnumber}{Optional. TFL number.}

\item{pname}{Optional. Program name.}

\item{oid}{Optional. Object ID.}

\item{add_footr_tstamp}{If TRUE, add timestamp and source information as the
last line of the footnotes. Default is TRUE.}
}
\value{
A named list of non-missing footnote fields.
}
\description{
Retrieves footnote fields (FOOT1, FOOT2, etc.) from the metadata.
}
