% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_pgmname}
\alias{get_pgmname}
\title{Extract Program Name Metadata}
\usage{
get_pgmname(df, tnumber = NULL, pname = NULL, oid = NULL)
}
\arguments{
\item{df}{A data frame or tibble containing title and footnote metadata.}

\item{tnumber}{Optional TFL number.}

\item{pname}{Optional program name.}

\item{oid}{Optional object ID.}
}
\value{
A named list of non-missing program name field.
}
\description{
Retrieves program name field (PGMNAME) from metadata for a specified
program name or TFL number.
}
