% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_pgmname}
\alias{get_pgmname}
\title{Extract Program Name Metadata}
\usage{
get_pgmname(df = NULL, type = NULL, tnumber = NULL, pname = NULL, oid = NULL)
}
\arguments{
\item{df}{A data frame of title and footnote metadata.}

\item{type}{Optional. A character string indicating the TFL type.}

\item{tnumber}{Optional. TFL number.}

\item{pname}{Optional. Program name.}

\item{oid}{Optional. Object ID.}
}
\value{
A named list containing the program name field.
}
\description{
Retrieves program name field (PGMNAME) from the metadata.
}
