% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_pop}
\alias{get_pop}
\title{Extract Population Metadata}
\usage{
get_pop(df = NULL, type = NULL, tnumber = NULL, pname = NULL, oid = NULL)
}
\arguments{
\item{df}{A data frame or tibble of title and footnote metadata.}

\item{type}{Optional. A character string indicating the TFL type.}

\item{tnumber}{Optional. TFL number.}

\item{pname}{Optional. Program name.}

\item{oid}{Optional. Object ID.}
}
\value{
A named list containing the population field.
}
\description{
Retrieves population field from the metadata.
}
