% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_source}
\alias{get_source}
\title{Extract Source Metadata}
\usage{
get_source(df = NULL, type = NULL, tnumber = NULL, pname = NULL, oid = NULL)
}
\arguments{
\item{df}{A data frame of title and footnote metadata.}

\item{type}{Optional. A character string indicating the TFL type.}

\item{tnumber}{Optional. TFL number.}

\item{pname}{Optional. Program name.}

\item{oid}{Optional. Object ID.}
}
\value{
A named list of non-missing source fields.
}
\description{
Retrieves source fields (e.g., SOURCE1) from the metadata.
}
