% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils2.R
\name{get_source}
\alias{get_source}
\title{Return data source}
\usage{
get_source(filename, type = NA, tnumber = NA, pname = NA, oid = NA)
}
\arguments{
\item{filename}{Full filename including folder path and file extension.}

\item{type}{optional TFL type}

\item{tnumber}{TFL number, used to select proper titles and footnotes. This can be with or without TFL type.If pname parameter is not given, tnumber must not be NULL.}

\item{pname}{The program name used to select proper title entry. If this parameter is given (not NULL), it has precedence over TFL number for selection.}

\item{oid}{Optional parameter}
}
\description{
This function returns data source(s) for printing at the bottom of TFLs
}
\details{
Calls other utils functions.
}
