% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_timestamp}
\alias{get_timestamp}
\title{Generate Runtime Timestamp Reference}
\usage{
get_timestamp(pname = NULL)
}
\arguments{
\item{pname}{Optional. A program name to include in the timestamp message.}
}
\value{
A character string with the timestamp and source filename.
}
\description{
Generates a string indicating the time and source file from which the report was generated.
}
