% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_footer.R
\name{get_title}
\alias{get_title}
\title{Generate headers - titles and population}
\usage{
get_title(filename, type = NA, tnumber = NA, pname = NA, oid = NA)
}
\arguments{
\item{filename}{Full filename including folder path and file extension.}

\item{type}{optional TFL type}

\item{tnumber}{TFL number, used to select proper titles and footnotes. This can be with or without TFL type.If pname parameter is not given, tnumber must not be NULL.}

\item{pname}{The program name used to select proper title entry. If this parameter is given (not NULL), it has precedence over TFL number for selection.}

\item{oid}{Optional parameter}
}
\description{
This function reads headers for tables and listings which will be generated using {flextable} package.
}
\details{
Calls other utils functions. The header should be unlisted.
}
