% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_title}
\alias{get_title}
\title{Extract Titles and Subtitle Metadata}
\usage{
get_title(df = NULL, type = NULL, tnumber = NULL, pname = NULL, oid = NULL)
}
\arguments{
\item{df}{A data frame or tibble of title and footnote metadata.}

\item{type}{Optional. A character string indicating the TFL type.}

\item{tnumber}{Optional. TFL number.}

\item{pname}{Optional. Program name.}

\item{oid}{Optional. Object ID.}
}
\value{
A named list of non-missing title fields.
}
\description{
Retrieves title-related fields (TTL1, TTL2, etc.) and population from the metadata.
}
