% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ulheader}
\alias{get_ulheader}
\title{Get upper left headers}
\usage{
get_ulheader(df, by_list = TRUE)
}
\arguments{
\item{df}{a full filename including folder path.}

\item{by_list}{optional return type. Default is a list.}

\item{sheetname}{optional sheetname. Default is "header"}
}
\description{
Get upper left headers
}
\details{
Generate a list of upper left header. If by_list = FALSE, the function will return a string vector.
}
