% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.R
\name{get_ulheader}
\alias{get_ulheader}
\title{Get upper left headers}
\usage{
get_ulheader(filename, sheetname = "header", by_list = TRUE)
}
\arguments{
\item{filename}{a full filename including folder path.}

\item{sheetname}{optional sheetname. Default is "header"}

\item{by_list}{optional return type. Default is a list.}
}
\description{
Get upper left headers
}
\details{
Generate a list of upper left header. If by_list = FALSE, the function will return a string vector.
}
