% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_urheader}
\alias{get_urheader}
\title{Get upper right headers}
\usage{
get_urheader(df, by_list = TRUE)
}
\arguments{
\item{df}{a full filename including folder path.}

\item{by_list}{optional return type. Default is a list.}

\item{sheetname}{optional sheetname. Default is "header"}
}
\description{
Get upper right headers
}
\details{
Generate a list of upper right header. If by_list = FALSE, the function will return a string vector.
}
