% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_footer}
\alias{read_footer}
\title{Read Titles and Footnotes Excel File}
\usage{
read_footer(filename = filename, sheetname = NULL)
}
\arguments{
\item{filename}{a full filename including folder path.}

\item{sheetname}{optional sheetname}
}
\description{
This function reads headers and footers from an excel spreadsheet.
}
\details{
The title and footnotes are to be maintained and managed in the excel spreadsheet. This file should have the following column names in order for the program
to pull appropriate header or footnotes:\cr
\cr "TYPE", "PGMNAME", "OID", "TTL1", "SOURCE", "BYLINE1", "FOOT1" \cr
Filename should have the complete folder path and correct file extension. \cr

For an example, see \code{vignette("use_flextable", package = "TFootr")}.
}
\examples{

titles_footnotes <- read_footer("path/to/your/titles.xls")


}
\seealso{
\link{get_footnote}, \link{select_row_header}
}
