% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_footer}
\alias{read_footer}
\title{Read the Excel file of header and footer}
\usage{
read_footer(filename = filename)
}
\description{
This function reads headers and footers from an excel spreadsheet.
}
\details{
The title and footnotes are to be maintained and managed in the excel spreadsheet. This file should have the following column names in order for the program
to pull appropriate header or footnotes:\cr
\tab "TYPE", "PGMNAME", "OID", "TTL1", "SOURCE", "BYLINE1", "FOOT1" \cr
Filename should have the complete folder path and correct file extension. \cr

For an example, see \code{vignette("Use_Flextable", package = "headr")}.
}
\examples{


# (to use |> version 4.1.0 of R is required, for lower versions we recommend \%>\% from magrittr)
library(magrittr)

library(flextable)



}
\seealso{
\link{read_header}
}
