% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_header}
\alias{read_header}
\title{Read headers from Excel File}
\usage{
read_header(filename = filename, sheetname = "header")
}
\arguments{
\item{filename}{a full filename including folder path.}

\item{sheetname}{optional sheetname}
}
\description{
This function reads headers from an excel spreadsheet.
The headers are retrieved from the same excel file but in "header" sheet.
If the sheetname is not standard, the name should be provided.
}
\details{
Both left and right headers are in a separate sheet of the same file as the title and footnotes.\cr
This file should have the following column names in order for the program
to pull appropriate headers:\cr
\cr "UL1", "UL2", "UL3", "UR1", "UR2", "UR3", even though two are needed for each side. \cr
All other columns will be discarded.Only the first row will be read.
Filename should have the complete folder path and correct file extension. \cr

For an example, see \code{vignette("use_flextable", package = "TFootr")}.
}
\examples{

headers <- read_header("path/to/your/titles.xls")


}
\seealso{
\link{read_footer}
}
