% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_tfile}
\alias{read_tfile}
\title{Read Titles and Footnotes Excel File}
\usage{
read_tfile(filename = NULL, sheetname = NULL, ...)
}
\arguments{
\item{filename}{a full filename including folder path.}

\item{sheetname}{default sheetname for title & footnote is "title_footnote"; default sheetname for header part is "header". If not given, the function will read the first sheet.}
}
\description{
A wrapper function to read title & footnote excel spreadsheet file
}
\details{
The title and footnotes are to be maintained in the excel spreadsheet. This file should have the following column names in order for the program
to pull appropriate header or footnotes:\cr
\cr "PGMNAME", "TTL1", "SOURCE", "FOOT1" \cr
Filename can be an object returned by file.path() function  \cr
}
\examples{

}
\seealso{
\link{select_row}
}
