% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_row.R
\name{select_row_header}
\alias{select_row_header}
\title{Select header and footer based on program name or TFL number}
\usage{
select_row_header(
  filename,
  type = NULL,
  tnumber = NULL,
  pname = NULL,
  oid = NULL
)
}
\arguments{
\item{type}{optional TFL type}

\item{tnumber}{TFL number, used to select proper titles and footnotes. This can be with or without TFL type.If pname parameter is not given, tnumber must not be NA.}

\item{pname}{The program name used to select proper title entry. If this parameter is given (not NA), it has precedence over TFL number for selection.}

\item{oid}{Optional parameter}

\item{df}{A dataframe or list of overall title and footnote}
}
\description{
Select header and footer based on program name or TFL number
}
\details{
Additional details...
}
