% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_row.R
\name{select_with_name}
\alias{select_with_name}
\title{Select a Metadata Entry by Program Name and Optional OID}
\usage{
select_with_name(df = list(), pname = "", oid = "")
}
\arguments{
\item{df}{A data frame containing the metadata.}

\item{pname}{A character string specifying the program name.}

\item{oid}{Optional. A character string specifying the object ID.}
}
\value{
A single-row data frame (list) with matching metadata.
}
\description{
Filters the metadata for a unique match based on program name and optional object ID (OID).
}
