% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_with_number}
\alias{select_with_number}
\title{Select header and footer based on TFL number}
\usage{
select_with_number(df = list(), type = NA, tnumber = "")
}
\description{
This function select titles and footnotes based on TFL number
}
\details{
Each TFL (table, listing, or figure) should have a type and TFL number in the header_footer Excel spreadsheet. This information can be in two separate columns (TYPE & TTL1)
or combined in TTL1 column. However, if both type and TFL number are already combined in column TTL1, there should be a space to separate them.
Additional details...
}
