% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tflmetaR.R
\name{tflmetaR}
\alias{tflmetaR}
\title{Get title and footnotes from an Excel file}
\usage{
tflmetaR(
  xlfile,
  sheet_name,
  by_column = "PGMNAME",
  by_value,
  select_type = NULL,
  add_footr_tstamp = TRUE,
  oid = NULL
)
}
\arguments{
\item{xlfile}{Filename including folder path and file extension (xls/xlsx).}

\item{sheet_name}{Sheet to read.}

\item{by_column}{Column name that is used to filter proper titles and
footnotes. Default is "PGMNAME".}

\item{by_value}{Column value that is used to filter proper titles and
footnotes.}

\item{select_type}{Either NULL to return the entire selected row, "title" to
return columns starting with "TTL", "footr" to return columns starting with
"FOOT", or a specific column name to return that particular column. The
default is NULL.}

\item{add_footr_tstamp}{If TRUE, add timestamp and source information as the
last line of the footnotes. It is only applied when select_type = "footr"}

\item{oid}{If not NULL, oid is used to filter the appropriate titles and
footnotes. The default value is NULL.}
}
\value{
Returns a data frame
}
\description{
Get title and footnotes from an Excel file
}
\examples{
\dontrun{
(1) To get titles:  tflmetaR(filename, "Sheet1", by_value="t_dm", select_type="title")
(2) To get footers: tflmetaR(filename, "Sheet1", by_value="t_dm", select_type="footr")
(3) To get a specific cell value, i.e., SOURCE:
      tflmetaR(filename, "Sheet1", by_value="t_dm", select_type="source")
(4) To get the whole row: tflmetaR(filename, "Sheet1", by_value="t_dm")
}
}
